+++ CHANGELOG.md
@@ -6,7 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased 0.8.0 - 
 
-## Changed
+### Added
+- Deactivate theme if minimum required WordPress and PHP versions not met
+
+### Changed
 - Move `isPagelike()` from posts utility to layout utility.
 
 ## 0.7.1 - 2018-06-24